+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ CHANGELOG
 * doc-fix: Add information on necessary AWS permissions
 * bug-fix: Remove PyYAML to let docker-compose install the right version
 * feature: Update TensorFlow latest version to 1.12
+* enhancement: Add Model.transformer()
+
 
 1.16.3
 ======